+++ src/elite/perf/EliteSharpLib.Benchmarks/SunBenchmarks.cs
@@ -28,7 +28,8 @@ public class SunBenchmarks : IDisposable
         Useful.Abstraction.ScreenManager<Views.Screen, Views.IView> views = new(keyboard);
         GameState gameState = new(views);
         _graphics = SoftwareGraphics.Create(ScreenWidth, ScreenHeight, (_) => { }, assetLocator);
-        EliteDraw draw = new(gameState, _graphics, assetLocator);
+        ShipRenderer shipRenderer = new(gameState, _graphics, assetLocator);
+        EliteDraw draw = new(gameState, _graphics, assetLocator, shipRenderer);
         _gradientSun = new(draw);
         _solidSun = new(draw);
     }